Split TorrentHandle interface and implementation
