+++ src/base/bittorrent/session.h
@@ -93,6 +93,7 @@ namespace BitTorrent
     class InfoHash;
     class MagnetUri;
     class TorrentHandle;
+    class TorrentHandleImpl;
     class Tracker;
     class TrackerEntry;
     struct CreateTorrentParams;
@@ -411,7 +412,7 @@ namespace BitTorrent
 
         void startUpTorrents();
         TorrentHandle *findTorrent(const InfoHash &hash) const;
-        QHash<InfoHash, TorrentHandle *> torrents() const;
+        QVector<TorrentHandle *> torrents() const;
         bool hasActiveTorrents() const;
         bool hasUnfinishedTorrents() const;
         bool hasRunningSeed() const;
@@ -440,31 +441,31 @@ namespace BitTorrent
         void bottomTorrentsQueuePos(const QVector<InfoHash> &hashes);
 
         // TorrentHandle interface
-        void handleTorrentSaveResumeDataRequested(const TorrentHandle *torrent);
-        void handleTorrentShareLimitChanged(TorrentHandle *const torrent);
-        void handleTorrentNameChanged(TorrentHandle *const torrent);
-        void handleTorrentSavePathChanged(TorrentHandle *const torrent);
-        void handleTorrentCategoryChanged(TorrentHandle *const torrent, const QString &oldCategory);
-        void handleTorrentTagAdded(TorrentHandle *const torrent, const QString &tag);
-        void handleTorrentTagRemoved(TorrentHandle *const torrent, const QString &tag);
-        void handleTorrentSavingModeChanged(TorrentHandle *const torrent);
-        void handleTorrentMetadataReceived(TorrentHandle *const torrent);
-        void handleTorrentPaused(TorrentHandle *const torrent);
-        void handleTorrentResumed(TorrentHandle *const torrent);
-        void handleTorrentChecked(TorrentHandle *const torrent);
-        void handleTorrentFinished(TorrentHandle *const torrent);
-        void handleTorrentTrackersAdded(TorrentHandle *const torrent, const QVector<TrackerEntry> &newTrackers);
-        void handleTorrentTrackersRemoved(TorrentHandle *const torrent, const QVector<TrackerEntry> &deletedTrackers);
-        void handleTorrentTrackersChanged(TorrentHandle *const torrent);
-        void handleTorrentUrlSeedsAdded(TorrentHandle *const torrent, const QVector<QUrl> &newUrlSeeds);
-        void handleTorrentUrlSeedsRemoved(TorrentHandle *const torrent, const QVector<QUrl> &urlSeeds);
-        void handleTorrentResumeDataReady(TorrentHandle *const torrent, const std::shared_ptr<lt::entry> &data);
-        void handleTorrentResumeDataFailed(TorrentHandle *const torrent);
-        void handleTorrentTrackerReply(TorrentHandle *const torrent, const QString &trackerUrl);
-        void handleTorrentTrackerWarning(TorrentHandle *const torrent, const QString &trackerUrl);
-        void handleTorrentTrackerError(TorrentHandle *const torrent, const QString &trackerUrl);
+        void handleTorrentSaveResumeDataRequested(const TorrentHandleImpl *torrent);
+        void handleTorrentShareLimitChanged(TorrentHandleImpl *const torrent);
+        void handleTorrentNameChanged(TorrentHandleImpl *const torrent);
+        void handleTorrentSavePathChanged(TorrentHandleImpl *const torrent);
+        void handleTorrentCategoryChanged(TorrentHandleImpl *const torrent, const QString &oldCategory);
+        void handleTorrentTagAdded(TorrentHandleImpl *const torrent, const QString &tag);
+        void handleTorrentTagRemoved(TorrentHandleImpl *const torrent, const QString &tag);
+        void handleTorrentSavingModeChanged(TorrentHandleImpl *const torrent);
+        void handleTorrentMetadataReceived(TorrentHandleImpl *const torrent);
+        void handleTorrentPaused(TorrentHandleImpl *const torrent);
+        void handleTorrentResumed(TorrentHandleImpl *const torrent);
+        void handleTorrentChecked(TorrentHandleImpl *const torrent);
+        void handleTorrentFinished(TorrentHandleImpl *const torrent);
+        void handleTorrentTrackersAdded(TorrentHandleImpl *const torrent, const QVector<TrackerEntry> &newTrackers);
+        void handleTorrentTrackersRemoved(TorrentHandleImpl *const torrent, const QVector<TrackerEntry> &deletedTrackers);
+        void handleTorrentTrackersChanged(TorrentHandleImpl *const torrent);
+        void handleTorrentUrlSeedsAdded(TorrentHandleImpl *const torrent, const QVector<QUrl> &newUrlSeeds);
+        void handleTorrentUrlSeedsRemoved(TorrentHandleImpl *const torrent, const QVector<QUrl> &urlSeeds);
+        void handleTorrentResumeDataReady(TorrentHandleImpl *const torrent, const std::shared_ptr<lt::entry> &data);
+        void handleTorrentResumeDataFailed(TorrentHandleImpl *const torrent);
+        void handleTorrentTrackerReply(TorrentHandleImpl *const torrent, const QString &trackerUrl);
+        void handleTorrentTrackerWarning(TorrentHandleImpl *const torrent, const QString &trackerUrl);
+        void handleTorrentTrackerError(TorrentHandleImpl *const torrent, const QString &trackerUrl);
 
-        bool addMoveTorrentStorageJob(TorrentHandle *torrent, const QString &newPath, MoveStorageMode mode);
+        bool addMoveTorrentStorageJob(TorrentHandleImpl *torrent, const QString &newPath, MoveStorageMode mode);
 
     signals:
         void addTorrentFailed(const QString &error);
@@ -521,7 +522,7 @@ namespace BitTorrent
     private:
         struct MoveStorageJob
         {
-            TorrentHandle *torrent;
+            TorrentHandleImpl *torrent;
             QString path;
             MoveStorageMode mode;
         };
@@ -571,7 +572,7 @@ namespace BitTorrent
         bool findIncompleteFiles(TorrentInfo &torrentInfo, QString &savePath) const;
 
         void updateSeedingLimitTimer();
-        void exportTorrentFile(TorrentHandle *const torrent, TorrentExportFolder folder = TorrentExportFolder::Regular);
+        void exportTorrentFile(const TorrentHandle *torrent, TorrentExportFolder folder = TorrentExportFolder::Regular);
 
         void handleAlert(const lt::alert *a);
         void dispatchTorrentAlert(const lt::alert *a);
@@ -733,7 +734,7 @@ namespace BitTorrent
         ResumeDataSavingManager *m_resumeDataSavingManager = nullptr;
 
         QHash<InfoHash, TorrentInfo> m_loadedMetadata;
-        QHash<InfoHash, TorrentHandle *> m_torrents;
+        QHash<InfoHash, TorrentHandleImpl *> m_torrents;
         QHash<InfoHash, CreateTorrentParams> m_addingTorrents;
         QHash<QString, AddTorrentParams> m_downloadedTorrents;
         QHash<InfoHash, RemovingTorrentData> m_removingTorrents;
