+++ src/base/bittorrent/torrenthandle.h
@@ -27,83 +27,29 @@
  * exception statement from your version.
  */
 
-#ifndef BITTORRENT_TORRENTHANDLE_H
-#define BITTORRENT_TORRENTHANDLE_H
+#pragma once
 
-#include <functional>
-
-#include <libtorrent/fwd.hpp>
-#include <libtorrent/torrent_handle.hpp>
-#include <libtorrent/torrent_status.hpp>
-
-#include <QDateTime>
 #include <QHash>
-#include <QObject>
-#include <QQueue>
+#include <QMetaType>
 #include <QSet>
 #include <QString>
 #include <QVector>
-
-#include "private/speedmonitor.h"
-#include "infohash.h"
-#include "torrentinfo.h"
-
-extern const QString QB_EXT;
 
 class QBitArray;
 class QDateTime;
 class QStringList;
 class QUrl;
 
+extern const QString QB_EXT;
+
 namespace BitTorrent
 {
     enum class DownloadPriority;
+    class InfoHash;
     class PeerInfo;
-    class Session;
+    class TorrentInfo;
     class TrackerEntry;
-    struct AddTorrentParams;
     struct PeerAddress;
-
-    struct CreateTorrentParams
-    {
-        bool restored; // is existing torrent job?
-        // for both new and restored torrents
-        QString name;
-        QString category;
-        QSet<QString> tags;
-        QString savePath;
-        bool disableTempPath;
-        bool sequential;
-        bool firstLastPiecePriority;
-        bool hasSeedStatus;
-        bool skipChecking;
-        bool hasRootFolder;
-        bool forced;
-        bool paused;
-        int uploadLimit;
-        int downloadLimit;
-        // for new torrents
-        QVector<DownloadPriority> filePriorities;
-        QDateTime addedTime;
-        // for restored torrents
-        qreal ratioLimit;
-        int seedingTimeLimit;
-
-        CreateTorrentParams();
-        explicit CreateTorrentParams(const AddTorrentParams &params);
-    };
-
-    struct TrackerInfo
-    {
-        QString lastMessage;
-        int numPeers = 0;
-    };
-
-    enum class MoveStorageMode
-    {
-        KeepExistingFiles,
-        Overwrite
-    };
 
     enum class TorrentState
     {
@@ -135,13 +81,16 @@ namespace BitTorrent
         Error
     };
 
+    struct TrackerInfo
+    {
+        QString lastMessage;
+        int numPeers = 0;
+    };
+
     uint qHash(TorrentState key, uint seed);
 
-    class TorrentHandle : public QObject
+    class TorrentHandle
     {
-        Q_DISABLE_COPY(TorrentHandle)
-        Q_DECLARE_TR_FUNCTIONS(BitTorrent::TorrentHandle)
-
     public:
         static const qreal USE_GLOBAL_RATIO;
         static const qreal NO_RATIO_LIMIT;
@@ -152,24 +101,21 @@ namespace BitTorrent
         static const qreal MAX_RATIO;
         static const int MAX_SEEDING_TIME;
 
-        TorrentHandle(Session *session, const lt::torrent_handle &nativeHandle,
-                          const CreateTorrentParams &params);
-        ~TorrentHandle();
+        virtual ~TorrentHandle() = default;
 
-        bool isValid() const;
-        InfoHash hash() const;
-        QString name() const;
-        QDateTime creationDate() const;
-        QString creator() const;
-        QString comment() const;
-        bool isPrivate() const;
-        qlonglong totalSize() const;
-        qlonglong wantedSize() const;
-        qlonglong completedSize() const;
-        qlonglong incompletedSize() const;
-        qlonglong pieceLength() const;
-        qlonglong wastedSize() const;
-        QString currentTracker() const;
+        virtual InfoHash hash() const = 0;
+        virtual QString name() const = 0;
+        virtual QDateTime creationDate() const = 0;
+        virtual QString creator() const = 0;
+        virtual QString comment() const = 0;
+        virtual bool isPrivate() const = 0;
+        virtual qlonglong totalSize() const = 0;
+        virtual qlonglong wantedSize() const = 0;
+        virtual qlonglong completedSize() const = 0;
+        virtual qlonglong incompletedSize() const = 0;
+        virtual qlonglong pieceLength() const = 0;
+        virtual qlonglong wastedSize() const = 0;
+        virtual QString currentTracker() const = 0;
 
         // 1. savePath() - the path where all the files and subfolders of torrent are stored (as always).
         // 2. rootPath() - absolute path of torrent file tree (save path + first item from 1st torrent file path).
@@ -214,237 +160,140 @@ namespace BitTorrent
         // | B | /home/user/torrents/torrentB | /home/user/torrents/torrentB/subdir1/file1 |
         // | C | /home/user/torrents/file1    | /home/user/torrents/file1                  |
 
-        QString savePath(bool actual = false) const;
-        QString rootPath(bool actual = false) const;
-        QString contentPath(bool actual = false) const;
+        virtual QString savePath(bool actual = false) const = 0;
+        virtual QString rootPath(bool actual = false) const = 0;
+        virtual QString contentPath(bool actual = false) const = 0;
 
-        bool useTempPath() const;
+        virtual bool useTempPath() const = 0;
 
-        bool isAutoTMMEnabled() const;
-        void setAutoTMMEnabled(bool enabled);
-        QString category() const;
-        bool belongsToCategory(const QString &category) const;
-        bool setCategory(const QString &category);
+        virtual bool isAutoTMMEnabled() const = 0;
+        virtual void setAutoTMMEnabled(bool enabled) = 0;
+        virtual QString category() const = 0;
+        virtual bool belongsToCategory(const QString &category) const = 0;
+        virtual bool setCategory(const QString &category) = 0;
 
-        QSet<QString> tags() const;
-        bool hasTag(const QString &tag) const;
-        bool addTag(const QString &tag);
-        bool removeTag(const QString &tag);
-        void removeAllTags();
+        virtual QSet<QString> tags() const = 0;
+        virtual bool hasTag(const QString &tag) const = 0;
+        virtual bool addTag(const QString &tag) = 0;
+        virtual bool removeTag(const QString &tag) = 0;
+        virtual void removeAllTags() = 0;
 
-        bool hasRootFolder() const;
+        virtual bool hasRootFolder() const = 0;
 
-        int filesCount() const;
-        int piecesCount() const;
-        int piecesHave() const;
-        qreal progress() const;
-        QDateTime addedTime() const;
-        qreal ratioLimit() const;
-        int seedingTimeLimit() const;
+        virtual int filesCount() const = 0;
+        virtual int piecesCount() const = 0;
+        virtual int piecesHave() const = 0;
+        virtual qreal progress() const = 0;
+        virtual QDateTime addedTime() const = 0;
+        virtual qreal ratioLimit() const = 0;
+        virtual int seedingTimeLimit() const = 0;
 
-        QString filePath(int index) const;
-        QString fileName(int index) const;
-        qlonglong fileSize(int index) const;
-        QStringList absoluteFilePaths() const;
-        QStringList absoluteFilePathsUnwanted() const;
-        QVector<DownloadPriority> filePriorities() const;
+        virtual QString filePath(int index) const = 0;
+        virtual QString fileName(int index) const = 0;
+        virtual qlonglong fileSize(int index) const = 0;
+        virtual QStringList absoluteFilePaths() const = 0;
+        virtual QStringList absoluteFilePathsUnwanted() const = 0;
+        virtual QVector<DownloadPriority> filePriorities() const = 0;
 
-        TorrentInfo info() const;
-        bool isSeed() const;
-        bool isPaused() const;
-        bool isResumed() const;
-        bool isQueued() const;
-        bool isForced() const;
-        bool isChecking() const;
-        bool isDownloading() const;
-        bool isUploading() const;
-        bool isCompleted() const;
-        bool isActive() const;
-        bool isInactive() const;
-        bool isErrored() const;
-        bool isSequentialDownload() const;
-        bool hasFirstLastPiecePriority() const;
-        TorrentState state() const;
-        bool hasMetadata() const;
-        bool hasMissingFiles() const;
-        bool hasError() const;
-        bool hasFilteredPieces() const;
-        int queuePosition() const;
-        QVector<TrackerEntry> trackers() const;
-        QHash<QString, TrackerInfo> trackerInfos() const;
-        QVector<QUrl> urlSeeds() const;
-        QString error() const;
-        qlonglong totalDownload() const;
-        qlonglong totalUpload() const;
-        qlonglong activeTime() const;
-        qlonglong finishedTime() const;
-        qlonglong seedingTime() const;
-        qlonglong eta() const;
-        QVector<qreal> filesProgress() const;
-        int seedsCount() const;
-        int peersCount() const;
-        int leechsCount() const;
-        int totalSeedsCount() const;
-        int totalPeersCount() const;
-        int totalLeechersCount() const;
-        int completeCount() const;
-        int incompleteCount() const;
-        QDateTime lastSeenComplete() const;
-        QDateTime completedTime() const;
-        qlonglong timeSinceUpload() const;
-        qlonglong timeSinceDownload() const;
-        qlonglong timeSinceActivity() const;
-        int downloadLimit() const;
-        int uploadLimit() const;
-        bool superSeeding() const;
-        QVector<PeerInfo> peers() const;
-        QBitArray pieces() const;
-        QBitArray downloadingPieces() const;
-        QVector<int> pieceAvailability() const;
-        qreal distributedCopies() const;
-        qreal maxRatio() const;
-        int maxSeedingTime() const;
-        qreal realRatio() const;
-        int uploadPayloadRate() const;
-        int downloadPayloadRate() const;
-        qlonglong totalPayloadUpload() const;
-        qlonglong totalPayloadDownload() const;
-        int connectionsCount() const;
-        int connectionsLimit() const;
-        qlonglong nextAnnounce() const;
-
-        void setName(const QString &name);
-        void setSequentialDownload(bool enable);
-        void toggleSequentialDownload();
-        void setFirstLastPiecePriority(bool enabled);
-        void toggleFirstLastPiecePriority();
-        void pause();
-        void resume(bool forced = false);
-        void move(QString path);
-        void forceReannounce(int index = -1);
-        void forceDHTAnnounce();
-        void forceRecheck();
-        void renameFile(int index, const QString &name);
-        void prioritizeFiles(const QVector<DownloadPriority> &priorities);
-        void setRatioLimit(qreal limit);
-        void setSeedingTimeLimit(int limit);
-        void setUploadLimit(int limit);
-        void setDownloadLimit(int limit);
-        void setSuperSeeding(bool enable);
-        void flushCache() const;
-        void addTrackers(const QVector<TrackerEntry> &trackers);
-        void replaceTrackers(const QVector<TrackerEntry> &trackers);
-        void addUrlSeeds(const QVector<QUrl> &urlSeeds);
-        void removeUrlSeeds(const QVector<QUrl> &urlSeeds);
-        bool connectPeer(const PeerAddress &peerAddress);
-
-        QString toMagnetUri() const;
-
-        bool needSaveResumeData() const;
-
-        // Session interface
-        lt::torrent_handle nativeHandle() const;
-
-        void handleAlert(const lt::alert *a);
-        void handleStateUpdate(const lt::torrent_status &nativeStatus);
-        void handleTempPathChanged();
-        void handleCategorySavePathChanged();
-        void handleAppendExtensionToggled();
-        void saveResumeData();
-        void handleStorageMoved(const QString &newPath, const QString &errorMessage);
-
+        virtual TorrentInfo info() const = 0;
+        virtual bool isSeed() const = 0;
+        virtual bool isPaused() const = 0;
+        virtual bool isResumed() const = 0;
+        virtual bool isQueued() const = 0;
+        virtual bool isForced() const = 0;
+        virtual bool isChecking() const = 0;
+        virtual bool isDownloading() const = 0;
+        virtual bool isUploading() const = 0;
+        virtual bool isCompleted() const = 0;
+        virtual bool isActive() const = 0;
+        virtual bool isInactive() const = 0;
+        virtual bool isErrored() const = 0;
+        virtual bool isSequentialDownload() const = 0;
+        virtual bool hasFirstLastPiecePriority() const = 0;
+        virtual TorrentState state() const = 0;
+        virtual bool hasMetadata() const = 0;
+        virtual bool hasMissingFiles() const = 0;
+        virtual bool hasError() const = 0;
+        virtual bool hasFilteredPieces() const = 0;
+        virtual int queuePosition() const = 0;
+        virtual QVector<TrackerEntry> trackers() const = 0;
+        virtual QHash<QString, TrackerInfo> trackerInfos() const = 0;
+        virtual QVector<QUrl> urlSeeds() const = 0;
+        virtual QString error() const = 0;
+        virtual qlonglong totalDownload() const = 0;
+        virtual qlonglong totalUpload() const = 0;
+        virtual qlonglong activeTime() const = 0;
+        virtual qlonglong finishedTime() const = 0;
+        virtual qlonglong seedingTime() const = 0;
+        virtual qlonglong eta() const = 0;
+        virtual QVector<qreal> filesProgress() const = 0;
+        virtual int seedsCount() const = 0;
+        virtual int peersCount() const = 0;
+        virtual int leechsCount() const = 0;
+        virtual int totalSeedsCount() const = 0;
+        virtual int totalPeersCount() const = 0;
+        virtual int totalLeechersCount() const = 0;
+        virtual int completeCount() const = 0;
+        virtual int incompleteCount() const = 0;
+        virtual QDateTime lastSeenComplete() const = 0;
+        virtual QDateTime completedTime() const = 0;
+        virtual qlonglong timeSinceUpload() const = 0;
+        virtual qlonglong timeSinceDownload() const = 0;
+        virtual qlonglong timeSinceActivity() const = 0;
+        virtual int downloadLimit() const = 0;
+        virtual int uploadLimit() const = 0;
+        virtual bool superSeeding() const = 0;
+        virtual QVector<PeerInfo> peers() const = 0;
+        virtual QBitArray pieces() const = 0;
+        virtual QBitArray downloadingPieces() const = 0;
+        virtual QVector<int> pieceAvailability() const = 0;
+        virtual qreal distributedCopies() const = 0;
+        virtual qreal maxRatio() const = 0;
+        virtual int maxSeedingTime() const = 0;
+        virtual qreal realRatio() const = 0;
+        virtual int uploadPayloadRate() const = 0;
+        virtual int downloadPayloadRate() const = 0;
+        virtual qlonglong totalPayloadUpload() const = 0;
+        virtual qlonglong totalPayloadDownload() const = 0;
+        virtual int connectionsCount() const = 0;
+        virtual int connectionsLimit() const = 0;
+        virtual qlonglong nextAnnounce() const = 0;
         /**
          * @brief fraction of file pieces that are available at least from one peer
          *
          * This is not the same as torrrent availability, it is just a fraction of pieces
          * that can be downloaded right now. It varies between 0 to 1.
          */
-        QVector<qreal> availableFileFractions() const;
+        virtual QVector<qreal> availableFileFractions() const = 0;
 
-    private:
-        typedef std::function<void ()> EventTrigger;
+        virtual void setName(const QString &name) = 0;
+        virtual void setSequentialDownload(bool enable) = 0;
+        virtual void setFirstLastPiecePriority(bool enabled) = 0;
+        virtual void pause() = 0;
+        virtual void resume(bool forced = false) = 0;
+        virtual void move(QString path) = 0;
+        virtual void forceReannounce(int index = -1) = 0;
+        virtual void forceDHTAnnounce() = 0;
+        virtual void forceRecheck() = 0;
+        virtual void renameFile(int index, const QString &name) = 0;
+        virtual void prioritizeFiles(const QVector<DownloadPriority> &priorities) = 0;
+        virtual void setRatioLimit(qreal limit) = 0;
+        virtual void setSeedingTimeLimit(int limit) = 0;
+        virtual void setUploadLimit(int limit) = 0;
+        virtual void setDownloadLimit(int limit) = 0;
+        virtual void setSuperSeeding(bool enable) = 0;
+        virtual void flushCache() const = 0;
+        virtual void addTrackers(const QVector<TrackerEntry> &trackers) = 0;
+        virtual void replaceTrackers(const QVector<TrackerEntry> &trackers) = 0;
+        virtual void addUrlSeeds(const QVector<QUrl> &urlSeeds) = 0;
+        virtual void removeUrlSeeds(const QVector<QUrl> &urlSeeds) = 0;
+        virtual bool connectPeer(const PeerAddress &peerAddress) = 0;
 
-#if (LIBTORRENT_VERSION_NUM < 10200)
-        using LTFileIndex = int;
-#else
-        using LTFileIndex = lt::file_index_t;
-#endif
+        virtual QString createMagnetURI() const = 0;
 
-        void updateStatus();
-        void updateStatus(const lt::torrent_status &nativeStatus);
-        void updateState();
-        void updateTorrentInfo();
-
-        void handleFastResumeRejectedAlert(const lt::fastresume_rejected_alert *p);
-        void handleFileCompletedAlert(const lt::file_completed_alert *p);
-        void handleFileRenamedAlert(const lt::file_renamed_alert *p);
-        void handleFileRenameFailedAlert(const lt::file_rename_failed_alert *p);
-        void handleMetadataReceivedAlert(const lt::metadata_received_alert *p);
-        void handlePerformanceAlert(const lt::performance_alert *p) const;
-        void handleSaveResumeDataAlert(const lt::save_resume_data_alert *p);
-        void handleSaveResumeDataFailedAlert(const lt::save_resume_data_failed_alert *p);
-        void handleTorrentCheckedAlert(const lt::torrent_checked_alert *p);
-        void handleTorrentFinishedAlert(const lt::torrent_finished_alert *p);
-        void handleTorrentPausedAlert(const lt::torrent_paused_alert *p);
-        void handleTorrentResumedAlert(const lt::torrent_resumed_alert *p);
-        void handleTrackerErrorAlert(const lt::tracker_error_alert *p);
-        void handleTrackerReplyAlert(const lt::tracker_reply_alert *p);
-        void handleTrackerWarningAlert(const lt::tracker_warning_alert *p);
-
-        void resume_impl(bool forced);
-        bool isMoveInProgress() const;
-        QString actualStorageLocation() const;
-        bool isAutoManaged() const;
-        void setAutoManaged(bool enable);
-
-        void adjustActualSavePath();
-        void adjustActualSavePath_impl();
-        void move_impl(QString path, MoveStorageMode mode);
-        void moveStorage(const QString &newPath, MoveStorageMode mode);
-        void manageIncompleteFiles();
-        void setFirstLastPiecePriorityImpl(bool enabled, const QVector<DownloadPriority> &updatedFilePrio = {});
-
-        Session *const m_session;
-        lt::torrent_handle m_nativeHandle;
-        lt::torrent_status m_nativeStatus;
-        TorrentState m_state;
-        TorrentInfo m_torrentInfo;
-        SpeedMonitor m_speedMonitor;
-
-        InfoHash m_hash;
-
-        bool m_storageIsMoving = false;
-        // m_moveFinishedTriggers is activated only when the following conditions are met:
-        // all file rename jobs complete, all file move jobs complete
-        QQueue<EventTrigger> m_moveFinishedTriggers;
-        int m_renameCount;
-
-        // Until libtorrent provide an "old_name" field in `file_renamed_alert`
-        // we will rely on this workaround to remove empty leftover folders
-        QHash<LTFileIndex, QVector<QString>> m_oldPath;
-
-        bool m_useAutoTMM;
-
-        // Persistent data
-        QString m_name;
-        QString m_savePath;
-        QString m_category;
-        QSet<QString> m_tags;
-        bool m_hasSeedStatus;
-        qreal m_ratioLimit;
-        int m_seedingTimeLimit;
-        bool m_tempPathDisabled;
-        bool m_fastresumeDataRejected;
-        bool m_hasMissingFiles;
-        bool m_hasRootFolder;
-        bool m_needsToSetFirstLastPiecePriority;
-
-        QHash<QString, TrackerInfo> m_trackerInfos;
-
-        bool m_unchecked = false;
+        void toggleSequentialDownload();
+        void toggleFirstLastPiecePriority();
     };
 }
 
 Q_DECLARE_METATYPE(BitTorrent::TorrentState)
-
-#endif // BITTORRENT_TORRENTHANDLE_H
