+++ src/webui/api/synccontroller.cpp
@@ -34,6 +34,7 @@
 #include <QMetaObject>
 #include <QThread>
 
+#include "base/bittorrent/infohash.h"
 #include "base/bittorrent/peeraddress.h"
 #include "base/bittorrent/peerinfo.h"
 #include "base/bittorrent/session.h"
